conditionally move the file, if it exists

+++ nightly_backup.py
@@ -61,15 +61,17 @@ def run_backup(site):
 
 	### Move protected files into their own folder. Mainly because the COMIT script crashes if it tries to read these files (it can't skip files that it sees but lacks permission to read)
 	for protected_file in site['protected_files']:
-		# get the relative path of the file based on the string
-		protected_path = os.path.dirname(protected_file)
-		protected_filename = os.path.basename(protected_file)
-		protected_path_full_origin = nightly_dir + '/web/' + protected_path + '/'
-		protected_path_full_destination = nightly_dir + '/protected/' + protected_path + '/'
-		#print(protected_path_full_origin + protected_file)
-		#print(protected_path_full_destination + protected_file)
-		os.system('sudo -u ' + site['linux_user'] + ' ' + 'mkdir -p "' + protected_path_full_destination + '"')
-		os.system('sudo -u ' + site['linux_user'] + ' ' + 'mv -f "' + protected_path_full_origin + protected_filename + '" "' + protected_path_full_destination + protected_filename + '"')
+		# if the file actually exists, move it outside the web folder
+		if path.isfile(nightly_dir + '/web/' + protected_file):
+			# get the relative path of the file based on the string
+			protected_path = os.path.dirname(protected_file)
+			protected_filename = os.path.basename(protected_file)
+			protected_path_full_origin = nightly_dir + '/web/' + protected_path + '/'
+			protected_path_full_destination = nightly_dir + '/protected/' + protected_path + '/'
+			#print(protected_path_full_origin + protected_file)
+			#print(protected_path_full_destination + protected_file)
+			os.system('sudo -u ' + site['linux_user'] + ' ' + 'mkdir -p "' + protected_path_full_destination + '"')
+			os.system('sudo -u ' + site['linux_user'] + ' ' + 'mv -f "' + protected_path_full_origin + protected_filename + '" "' + protected_path_full_destination + protected_filename + '"')
 
 	### Backup COM Production database
 	# make sure there is no space between -p and the double quote
